% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{chol_downdate}
\alias{chol_downdate}
\title{Rank-one Downdate of Cholesky Decomposition}
\usage{
chol_downdate(L, u)
}
\arguments{
\item{L}{A lower triangular matrix. Strictly upper diagonal part is not referenced.}

\item{u}{A vector with with length matching with the dimensions of L.}
}
\value{
Updated L.
}
\description{
Given the lower triangular matrix L obtained from the Cholesky decomposition of A,
function \code{chol_downdate} updates L such that it corresponds to the decomposition of A - u*u'
(if such decomposition exists).
}
\note{
The function does not check that the resulting matrix is positive semidefinite.
}

