% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_S}
\alias{update_S}
\title{Update the Proposal of RAM Algorithm}
\usage{
update_S(S, u, current, n, target = 0.234, gamma = 2/3)
}
\arguments{
\item{S}{A lower triangular matrix corresponding to the Cholesky decomposition.}

\item{u}{A vector with with length matching with the dimensions of L.}

\item{current}{The current acceptance probability.}

\item{n}{Scaling parameter corresponding to the current iteration number.}

\item{target}{The target acceptance rate. Default is 0.234.}

\item{gamma}{Scaling parameter. Default is 2/3.}
}
\value{
Updated S.
}
\description{
Given the lower triangular matrix L obtained from the Cholesky decomposition of A,
function chol_downdate updates L such that it corresponds to the decomposition of A - u*u'
(if such decomposition exists).
}
\note{
The function does not check that the resulting matrix is positive semidefinite.
}

